Add Python and NodeJS to Programs
I also separated the 'programs' configuration to separate files
depending on usage. For now, I'm ok with Personal and Development, but I
might create more divisions later.

Most changes are because of the change in the folder structure, where I
get rid of the 'modules/dotfiles' and moved everything one folder back.